notes roll, new baselines, nixed source_file_mtime
This patch adds the concept of a notes roll, in which all published notes
appear.  Obviously, this has some scaling challenges, but I am kicking
that problem down the road for now.   The solution to this likely involves
generating JSON files with notes data that is consumed by the front end.

Also, I removed the concept of source_file_mtime as an attribute of Post
and Notes.  In the case of Post, the source can be rewritten and it is the
new mtime that should be recorded in a memory.  In any case, caching
this mtime value provides little efficiency, but introduces subtle bugs.

Plerd::publish_all is now a little more careful in what it chooses to
regenerate.  For example, if there are no new posts, but there are new
notes, all the views that use notes are regenerated, but those that solely
depend on posts are not.