add should_publish_THING()
This patch adds should_publish_note() and should_publish_post() to Plerd.

These methods determine whether the object in question should be published
or republished or declined.

This logic was previously in publish_all() and it was needlessly complex.

This change appears to pass all the existing tests and it is far easier to
read and maintain.

The root cause of the problem was premature optimization (again).