use site_description to control the about section
This patch adds support for a site_description parameter in the plerd
configuration file.  This is markdown that is converted into HTML for
the About this Site section of the side bar.

YAML has a convenient syntax for blocks of text.   Anything requiring
more formatting or complexity requires changing the site-specific
template.

I think this is a reasonable approach.