notes use source filen name as title; no HTML in notes
This patch changes plerd to prefer the name of the source file for the
title of the note.  If that is unavailble, it falls back to use the
content of the note.  Failing even that, the name of the published
file (which definitely will exist) is used.

*incompatiable change*

Notes aggressively attempt to find URLs and hyperlink them.  If you
want to post a gif, do not using an IMG tag.  People will have to
click through to the image. Life is hard.